sweetalert

+++ javascript/_modal.js
@@ -1,4 +1,3 @@
-
 import { eliminarArticuloCarrito } from "./_accionesCarrito.js";
 
 const modalContenedor = document.querySelector(".modal-contenedor");
@@ -23,6 +22,28 @@ modalCarrito.addEventListener("click", (e) => {
     e.stopPropagation();
 
     if (e.target.classList.contains('boton-eliminar')) {
-        eliminarArticuloCarrito(e.target.value)
+        Swal.fire({
+            title: 'Esta seguro?',
+            text: 'Va a eliminar el producto!',
+            icon: 'warning',
+            showCancelButton: true,
+            corfirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Eliminar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                eliminarArticuloCarrito(e.target.value)
+                Swal.fire(
+                    'Eliminado!',
+                    'El Producto ha sido eliminado.',
+                    'success'
+                )
+            }
+        });
+        
     };
 });
+
+   
+
